py_script/requirements.txt
pymysql
pyserial